requirements.txt
MetaTrader5>=5.0.45
pandas>=2.0.0
python-dotenv>=1.0.0
pytest>=7.4.0
